# Avoid recomputing `state.previous_epoch`