@paulhauner 2022-06-10
I'll reopen another PR with this change combined with others.